// apps/web/generated/api/models/ticketMessageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TicketMessageType } from './ticketMessageType';

export interface TicketMessageResponse {
  message_id: string;
  ticket_id: string;
  sender_user_id: string;
  message_type: TicketMessageType;
  body: string | null;
  is_internal: boolean;
  sent_at: string;
  attachment_file_asset_ids: string[];
}